Add correctness test and docu
Changed literal int size to i128 to fit the upper range of u64

+++ src/ast.rs
@@ -522,7 +522,7 @@ pub enum AstStatement {
     },
     // Literals
     LiteralInteger {
-        value: i64,
+        value: i128,
         location: SourceRange,
         id: AstId,
     },
